requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
decorator==4.3.0
enum34==1.1.6
firebase-admin==2.13.0
furl==2.0.0
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-firestore==0.30.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
grpcio==1.16.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.22
ipdb==0.11
ipython==5.8.0
ipython-genutils==0.2.0
jedi==0.13.1
lxml==4.2.5
msgpack==0.5.6
orderedmultidict==1.0
parsel==1.5.1
parso==0.3.1
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.3.0
PyHamcrest==1.9.0
pymongo==3.7.2
pyOpenSSL==18.0.0
python-decouple==3.1
pytz==2018.7
queuelib==1.5.0
recordtype==1.3
requests==2.20.1
rsa==4.0
scandir==1.9.0
Scrapy==1.5.1
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==18.9.0
urllib3==1.24.1
urlparse2==1.1.1
w3lib==1.19.0
wcwidth==0.1.7
zope.interface==4.6.0
